factor out mutations and make constants its own thing
