--- src/common/constants.json
@@ -1,3 +1,0 @@
-{
-	"MAX_TEAM_SIZE": 4
-}
