+++ src/common/constants.ts
@@ -1,0 +1,2 @@
+export const DEADLINE_TIMESTAMP = 1601679600000;
+export const MAX_TEAM_SIZE = 4;
